requirements.txt
aiofiles==25.1.0
anyio==4.12.0
asgiref==3.11.0
async-property==0.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
dj-database-url==3.1.0
Django==6.0
django-allauth==65.13.1
django-axes==8.1.0
django-cors-headers==4.9.0
django-oauth-toolkit==3.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==25.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
